requirements.txt
anthropic>=0.42.0
pandas>=2.2.0
matplotlib>=3.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.32.0
tabulate>=0.9.0
google-genai>=1.0.0
Pillow>=10.0.0
pyyaml>=6.0
apify-client>=1.0.0
beautifulsoup4>=4.12.0
